Initial separation of JSLint extension into generic linting manager (in core) and headless JSLint-specific 'linting provider' (in extension)
