+++ src/brackets.config.json
@@ -15,6 +15,7 @@
         "contributors_url"  : "https://api.github.com/repos/adobe/brackets/contributors",
         "extension_wiki_url": "https://github.com/adobe/brackets/wiki/Brackets-Extensions",
         "extension_registry": "https://s3.amazonaws.com/extend.brackets/registry.json",
-        "extension_url"     : "https://s3.amazonaws.com/extend.brackets/{0}/{0}-{1}.zip"
+        "extension_url"     : "https://s3.amazonaws.com/extend.brackets/{0}/{0}-{1}.zip",
+        "linting.enabled_by_default" : true
     }
 }
