+++ src/config.json
@@ -14,7 +14,8 @@
         "contributors_url": "https://api.github.com/repos/adobe/brackets/contributors",
         "extension_wiki_url": "https://github.com/adobe/brackets/wiki/Brackets-Extensions",
         "extension_registry": "https://s3.amazonaws.com/extend.brackets/registry.json",
-        "extension_url": "https://s3.amazonaws.com/extend.brackets/{0}/{0}-{1}.zip"
+        "extension_url": "https://s3.amazonaws.com/extend.brackets/{0}/{0}-{1}.zip",
+        "linting.enabled_by_default" : true
     },
     "name": "Brackets",
     "version": "0.28.0-0",
