--- src/extensions/default/JSLint/config.json
@@ -1,3 +1,0 @@
-{
-    "enabled_by_default": true
-}
