--- src/extensions/default/JSLint/jslint.css
@@ -1,31 +1,0 @@
-#status-indicators #jslint-status {
-    border: none;
-    cursor: default;	
-    width: 13px;
-}
-
-.jslint-disabled {
-    background: url(images/topcoat-inactive-15.svg) 9px 5px no-repeat;
-    pointer-events: none;
-}
-
-.jslint-errors {
-    background: url(images/topcoat-warning-15.svg) 9px 5px no-repeat;
-}
-
-.jslint-errors:hover {
-    background-color: rgba(0, 0, 0, 0.03);
-}
-
-.jslint-errors:active {
-    background-color: rgba(0, 0, 0, 0.05);
-}
-
-.jslint-valid {
-    background: url(images/topcoat-okay-15.svg) 9px 5px no-repeat;
-    pointer-events: none;
-}
-
-#jslint-results .line { /* To make the line number line up with editor line numbers */
-    text-align: right;
-}
