--- src/extensions/default/JSLint/keyboard.json
@@ -1,11 +1,0 @@
-{
-    "gotoFirstError":  [
-        {
-            "key": "F8"
-        },
-        {
-            "key": "Cmd-'",
-            "platform": "mac"
-        }
-    ]
-}
