+++ src/htmlContent/linting-panel.html
