+++ src/styles/brackets.less
@@ -904,10 +904,43 @@ a, img {
 }
 
 
-/* Turn on text selection for JSLint error messages */
-#jslint-results {
-    .user-select(text);
+/* Linting panel & status bar indicator */
+
+#status-indicators #jslint-status {
+    border: none;
+    cursor: default;
+    width: 13px;
 }
+
+.jslint-disabled {
+    background: url(images/topcoat-inactive-15.svg) 9px 5px no-repeat;
+    pointer-events: none;
+}
+
+.jslint-errors {
+    background: url(images/topcoat-warning-15.svg) 9px 5px no-repeat;
+}
+
+.jslint-errors:hover {
+    background-color: rgba(0, 0, 0, 0.03);
+}
+
+.jslint-errors:active {
+    background-color: rgba(0, 0, 0, 0.05);
+}
+
+.jslint-valid {
+    background: url(images/topcoat-okay-15.svg) 9px 5px no-repeat;
+    pointer-events: none;
+}
+
+#jslint-results {
+    .user-select(text);  // allow selecting error messages for easy web searching
+    .line {
+        text-align: right;  // make line number line up with editor line numbers
+    }
+}
+
 
 /* Line up label text and input text */
 label input {
